backend_v1/requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.4.2
# pydantic-settings==2.0.3
# sqlalchemy==2.0.23
# pymysql==1.1.0
# cryptography==41.0.7
# python-dotenv==1.0.0

# # AI/ML
# openai==1.3.5
# google-generativeai==0.3.0
# tiktoken==0.5.1
# langchain==0.0.340
# chromadb==0.4.18

# # 문서 처리
# pdfminer.six==20231228
# beautifulsoup4==4.12.2
# trafilatura==1.6.3
# requests==2.31.0

# # NLP
# nltk==3.8.1
# spacy==3.7.2

# # 개발/테스트
# pytest==7.4.3
# httpx==0.25.2
# trafilatura==1.6.3



fastapi
uvicorn
openai
google-generativeai
deep-translator
chromadb
langchain
tiktoken
pydantic
pydantic-settings
python-dotenv
sqlalchemy
PyMuPDF
pymysql
langchain_community
openai
langchain-text-splitters
google-generativeai
cryptography